requirements.txt
# Minimal requirements for the Telegram bot
aiogram==3.22.0
python-dotenv==1.0.1
